Renaming all variables...

+++ BEC-monopoles/utility.cpp
@@ -34,7 +34,7 @@ utility::~utility(){
     gsl_rng_free(r);
 }
 
-double utility::randnormed(int max){
+float utility::randnormed(int max){
     return gsl_rng_uniform (r) * max;
 }
 
@@ -42,31 +42,31 @@ int utility::randint(int max){
     return (int)gsl_rng_uniform_int(r,max);
 }
 
-double utility::randgaussian(double width){
+float utility::randgaussian(float width){
     return gsl_ran_gaussian_ziggurat(r, width);
 }
 
-double utility::vecavg(std::vector<double> v)
+float utility::vecavg(std::vector<float> v)
 {
-    double sum = accumulate(v.begin(), v.end(), 0.0);
-    double mean = sum / v.size();
+    float sum = accumulate(v.begin(), v.end(), 0.0);
+    float mean = sum / v.size();
     
     return mean;
 }
 
-double utility::vecstd(std::vector<double> v){
-    double mean = vecavg(v);
-    std::vector<double> diff(v.size());
-    std::transform(v.begin(), v.end(), diff.begin(),bind2nd(std::minus<double>(), mean));
-    double sq_sum = inner_product(diff.begin(), diff.end(), diff.begin(), 0.0);
-    double stdev = sqrt(sq_sum / v.size());
+float utility::vecstd(std::vector<float> v){
+    float mean = vecavg(v);
+    std::vector<float> diff(v.size());
+    std::transform(v.begin(), v.end(), diff.begin(),bind2nd(std::minus<float>(), mean));
+    float sq_sum = inner_product(diff.begin(), diff.end(), diff.begin(), 0.0);
+    float stdev = sqrt(sq_sum / v.size());
     
     return stdev;
 }
 
-std::vector<double> utility::vecadd(std::vector<double> a, std::vector<double> b){
-    std::vector<double> result;
-    std::transform(a.begin(), a.end(), b.begin(), std::back_inserter(result), [&](double l, double r)
+std::vector<float> utility::vecadd(std::vector<float> a, std::vector<float> b){
+    std::vector<float> result;
+    std::transform(a.begin(), a.end(), b.begin(), std::back_inserter(result), [&](float l, float r)
                    {
                        return r+l;
                    });
@@ -74,9 +74,9 @@ std::vector<double> utility::vecadd(std::vector<double> a, std::vector<double> b
     return result;
 }
 
-std::vector<double> utility::location(std::vector<double> bead, double boxsize){
+std::vector<float> utility::location(std::vector<float> bead, float boxsize){
     int ndim = (int)bead.size();
-    std::vector<double> loc;
+    std::vector<float> loc;
     for(int i = 0; i < ndim; i++){
         if(boxsize != -1)
             loc.push_back(pbc(bead[i],boxsize));
@@ -85,13 +85,13 @@ std::vector<double> utility::location(std::vector<double> bead, double boxsize){
     }
     return loc;
 }
-std::vector<double> utility::dist(std::vector<std::vector<double>> beads, double boxsize){
+std::vector<float> utility::dist(std::vector<std::vector<float>> beads, float boxsize){
     int ndim = (int)beads[0].size();
-    std::vector<double> dist;
-    std::vector<double> bead1 = location(beads[0],boxsize);
-    std::vector<double> bead2 = location(beads[1],boxsize);
+    std::vector<float> dist;
+    std::vector<float> bead1 = location(beads[0],boxsize);
+    std::vector<float> bead2 = location(beads[1],boxsize);
     for(int i = 0; i < ndim; i++){
-        double dimdist = bead2[i]-bead1[i];
+        float dimdist = bead2[i]-bead1[i];
         if(boxsize != -1)
             dimdist = pbc(dimdist+boxsize/2, boxsize)-boxsize/2;
         dist.push_back(dimdist);
@@ -99,19 +99,19 @@ std::vector<double> utility::dist(std::vector<std::vector<double>> beads, double
     return dist;
 }
 
-std::vector<double> utility::avedist(std::vector<std::vector<double>> beads, double boxsize){
-    std::vector<double> dstc = dist(beads, boxsize);
-    for(std::vector<double>::iterator it = dstc.begin(); it!= dstc.end(); it++){
+std::vector<float> utility::avedist(std::vector<std::vector<float>> beads, float boxsize){
+    std::vector<float> dstc = dist(beads, boxsize);
+    for(std::vector<float>::iterator it = dstc.begin(); it!= dstc.end(); it++){
         *it = *it/2 + beads[0][it-dstc.begin()];
     }
     return dstc;
 }
 
-double utility::pbc(double a, double b)
+float utility::pbc(float a, float b)
 {
     if(b < 0)
         return fmod(-a, -b);
-    double ret = fmod(a,b);
+    float ret = fmod(a,b);
     if(ret < 0)
         ret+=b;
     return ret;
